+++ Isotopos/src/main/java/isotopos/Isotopos/App.java
@@ -1,0 +1,13 @@
+package isotopos.Isotopos;
+
+/**
+ * Hello world!
+ *
+ */
+public class App 
+{
+    public static void main( String[] args )
+    {
+        System.out.println( "Hello World!" );
+    }
+}
