--- isotopos/Isotopos/Multas/.gitignore
@@ -1,1 +1,0 @@
-/target/
