--- isotopos/Isotopos/Multas/pom.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0"?>
-<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <groupId>edu.uclm.esi.iso2</groupId>
-    <artifactId>Isotopos</artifactId>
-    <version>0.0.1-SNAPSHOT</version>
-  </parent>
-  <groupId>edu.uclm.esi.iso2</groupId>
-  <artifactId>Multas</artifactId>
-  <version>0.0.1-SNAPSHOT</version>
-  <name>Multas</name>
-  <url>http://maven.apache.org</url>
-  <properties>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-  </properties>
-  <dependencies>
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <version>3.8.1</version>
-      <scope>test</scope>
-    </dependency>
-  </dependencies>
-</project>
