--- isotopos/Isotopos/Multas/src/main/java/org/Multas/App.java
@@ -1,13 +1,0 @@
-package org.Multas;
-
-/**
- * Mensaje de prueba
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
