--- isotopos/Isotopos/src/main/java/edu/uclm/esi/iso2/Isotopos/App.java
@@ -1,13 +1,0 @@
-package edu.uclm.esi.iso2.Isotopos;
-
-/**
- * Otra prueba.
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
